Move class MyCodeNode
